// src/types/formsTypes.tsx
export interface StatFormState {
    stat_id: string;
    name: string;
    description: string;
    formtype: string;
    // Add other fields as necessary
};

export interface StatToAbilityFormState {
    ability_id: string;
    name: string;
    formtype: string;
    // Add other fields as necessary
};